@@ -11,46 +11,47 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~7.2.12",
-    "@angular/cdk": "^7.3.7",
-    "@angular/common": "~7.2.12",
-    "@angular/compiler": "~7.2.12",
-    "@angular/core": "~7.2.12",
-    "@angular/forms": "~7.2.12",
-    "@angular/http": "^7.2.12",
-    "@angular/material": "^7.3.7",
-    "@angular/platform-browser": "~7.2.12",
-    "@angular/platform-browser-dynamic": "~7.2.12",
-    "@angular/router": "~7.2.12",
-    "@auth0/angular-jwt": "^2.1.0",
-    "@clr/angular": "^1.1.2",
-    "@clr/icons": "^1.1.2",
-    "@clr/ui": "^1.1.2",
-    "@types/chart.js": "^2.7.50",
-    "@webcomponents/custom-elements": "^1.2.2",
+    "@angular/animations": "^9.1.12",
+    "@angular/cdk": "^9.2.4",
+    "@angular/common": "^9.1.12",
+    "@angular/compiler": "~9.1.12",
+    "@angular/core": "^9.1.12",
+    "@angular/forms": "^9.1.12",
+    "@angular/material": "^9.2.4",
+    "@angular/platform-browser": "~9.1.12",
+    "@angular/platform-browser-dynamic": "~9.1.12",
+    "@angular/router": "^9.1.12",
+    "@auth0/angular-jwt": "^3.0.0",
+    "@clr/angular": "^3.1.4",
+    "@clr/core": "^3.0.0",
+    "@clr/icons": "^3.1.4",
+    "@clr/ui": "^3.1.4",
+    "@types/chart.js": "^2.8.4",
+    "@webcomponents/webcomponentsjs": "^2.0.0",
     "chart.js": "^2.8.0",
-    "core-js": "^3.0.1",
-    "rxjs": "~6.4.0",
-    "tslib": "^1.9.0",
-    "zone.js": "~0.9.0"
+    "core-js": "^3.2.1",
+    "rxjs": "~6.6.0",
+    "tslib": "^1.10.0",
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.13.8",
-    "@angular/cli": "~7.3.8",
-    "@angular/compiler-cli": "~7.2.12",
-    "@angular/language-service": "~7.2.12",
-    "@types/jasmine": "~3.3.12",
+    "@angular-devkit/build-angular": "~0.901.11",
+    "@angular/cli": "~9.1.11",
+    "@angular/compiler-cli": "^9.1.12",
+    "@angular/language-service": "^9.1.12",
+    "@types/jasmine": "^3.4.0",
     "@types/jasminewd2": "~2.0.3",
-    "@types/node": "^11.13.0",
-    "codelyzer": "^5.0.0",
+    "@types/node": "^12.11.1",
+    "codelyzer": "^5.1.2",
     "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.0.1",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma": "^5.1.0",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "^2.1.0",
     "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.0",
-    "protractor": "~6.0.0",
-    "ts-node": "~8.0.3"
+    "karma-jasmine-html-reporter": "^1.4.2",
+    "protractor": "^7.0.0",
+    "ts-node": "~8.3.0",
+    "typescript": "^3.8.3"
   }
 }
